Add email-password signup components for desktop and mobile
