+++ frontend/app/[lang]/signup/email-password/components/index.ts
@@ -1,0 +1,2 @@
+export * from './Desktop'
+export * from './Mobile'
